requirements.txt
# PriceOrders MVP - Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
supabase>=2.0.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0

# Fuzzy Matching
rapidfuzz>=3.5.0

# Frontend
streamlit>=1.28.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0

# ML Matching (serverless embeddings via OpenAI/OpenRouter)
openai>=1.0.0

# Telegram Bot
aiogram>=3.4.0
